add delete on every announcement

+++ anncs.php
@@ -41,9 +41,8 @@
 			<?php
 			$ID = $_GET['id'];
 			echo "<a href=\"./anncs/edit.php?id=" . $ID . "\"><button class=\"btn btn-default btn-event\">修改公告</button></a>";
-			echo "<a href=\"./anncs/delete.php?id=" . $ID . "\"><button class=\"btn btn-default btn-event\">刪除公告</button></a>";
-			/*<a href="signup.php"><button class="btn btn-default btn-event">修改公告</button></a>
-			<a href="delete.php?"><button class="btn btn-default btn-event">刪除公告</button></a>*/
+			echo "\r\n";
+			echo "<a href=\"./anncs/delete.php?id=" . $ID . "\" onclick=\"return confirm('使否確定要執行這個動作？');\"><button class=\"btn btn-default btn-event\">刪除公告</button></a>";
 			$Text_Name = "./AnncsText/" . $ID . ".txt";
 			$file = fopen($Text_Name, "r");
 			?>
